initiate split up of layer tests into submodules
