+++ deeplearning4j-modelimport/src/test/java/org/deeplearning4j/nn/modelimport/keras/layers/convolution/KerasConvolution2DTest.java
@@ -1,0 +1,136 @@
+/*-
+ *
+ *  * Copyright 2017 Skymind,Inc.
+ *  *
+ *  *    Licensed under the Apache License, Version 2.0 (the "License");
+ *  *    you may not use this file except in compliance with the License.
+ *  *    You may obtain a copy of the License at
+ *  *
+ *  *        http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  *    Unless required by applicable law or agreed to in writing, software
+ *  *    distributed under the License is distributed on an "AS IS" BASIS,
+ *  *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  *    See the License for the specific language governing permissions and
+ *  *    limitations under the License.
+ *
+ */
+package org.deeplearning4j.nn.modelimport.keras.layers.convolution;
+
+import org.deeplearning4j.nn.conf.ConvolutionMode;
+import org.deeplearning4j.nn.conf.layers.ConvolutionLayer;
+import org.deeplearning4j.nn.conf.layers.PoolingType;
+import org.deeplearning4j.nn.modelimport.keras.config.Keras1LayerConfiguration;
+import org.deeplearning4j.nn.modelimport.keras.config.Keras2LayerConfiguration;
+import org.deeplearning4j.nn.modelimport.keras.config.KerasLayerConfiguration;
+import org.deeplearning4j.nn.modelimport.keras.layers.convolutional.KerasConvolution2D;
+import org.deeplearning4j.nn.weights.WeightInit;
+import org.junit.Test;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+
+/**
+ * @author Max Pumperla
+ */
+public class KerasConvolution2DTest {
+
+    private final String ACTIVATION_KERAS = "linear";
+    private final String ACTIVATION_DL4J = "identity";
+    private final String LAYER_NAME = "test_layer";
+    private final String INIT_KERAS = "glorot_normal";
+    private final WeightInit INIT_DL4J = WeightInit.XAVIER;
+    private final double L1_REGULARIZATION = 0.01;
+    private final double L2_REGULARIZATION = 0.02;
+    private final double DROPOUT_KERAS = 0.3;
+    private final double DROPOUT_DL4J = 1 - DROPOUT_KERAS;
+    private final int[] KERNEL_SIZE = new int[]{1, 2};
+    private final int[] DILATION = new int[]{2, 2};
+    private final int[] INPUT_SHAPE = new int[]{100, 20};
+    private final int[] STRIDE = new int[]{3, 4};
+    private final PoolingType POOLING_TYPE = PoolingType.MAX;
+    private final int N_OUT = 13;
+    private final String BORDER_MODE_VALID = "valid";
+    private final int[] VALID_PADDING = new int[]{0, 0};
+
+    private Integer keras1 = 1;
+    private Integer keras2 = 2;
+    private Keras1LayerConfiguration conf1 = new Keras1LayerConfiguration();
+    private Keras2LayerConfiguration conf2 = new Keras2LayerConfiguration();
+
+
+    @Test
+    public void testConvolution2DLayer() throws Exception {
+        buildConvolution2DLayer(conf1, keras1, false);
+        buildConvolution2DLayer(conf2, keras2, false);
+        buildConvolution2DLayer(conf2, keras2, true);
+    }
+
+
+    public void buildConvolution2DLayer(KerasLayerConfiguration conf, Integer kerasVersion, boolean withDilation)
+            throws Exception {
+        Map<String, Object> layerConfig = new HashMap<String, Object>();
+        layerConfig.put(conf.getLAYER_FIELD_CLASS_NAME(), conf.getLAYER_CLASS_NAME_CONVOLUTION_2D());
+        Map<String, Object> config = new HashMap<String, Object>();
+        config.put(conf.getLAYER_FIELD_ACTIVATION(), ACTIVATION_KERAS); // keras linear -> dl4j identity
+        config.put(conf.getLAYER_FIELD_NAME(), LAYER_NAME);
+        if (kerasVersion == 1) {
+            config.put(conf.getLAYER_FIELD_INIT(), INIT_KERAS);
+        } else {
+            Map<String, Object> init = new HashMap<String, Object>();
+            init.put("class_name", conf.getINIT_GLOROT_NORMAL());
+            config.put(conf.getLAYER_FIELD_INIT(), init);
+        }
+        Map<String, Object> W_reg = new HashMap<String, Object>();
+        W_reg.put(conf.getREGULARIZATION_TYPE_L1(), L1_REGULARIZATION);
+        W_reg.put(conf.getREGULARIZATION_TYPE_L2(), L2_REGULARIZATION);
+        config.put(conf.getLAYER_FIELD_W_REGULARIZER(), W_reg);
+        config.put(conf.getLAYER_FIELD_DROPOUT(), DROPOUT_KERAS);
+        if (kerasVersion == 1) {
+            config.put(conf.getLAYER_FIELD_NB_ROW(), KERNEL_SIZE[0]);
+            config.put(conf.getLAYER_FIELD_NB_COL(), KERNEL_SIZE[1]);
+        } else {
+            ArrayList kernel = new ArrayList<Integer>() {{
+                for (int i : KERNEL_SIZE) add(i);
+            }};
+            config.put(conf.getLAYER_FIELD_KERNEL_SIZE(), kernel);
+        }
+        if (withDilation) {
+            ArrayList dilation = new ArrayList<Integer>() {{
+                for (int i : DILATION) add(i);
+            }};
+            config.put(conf.getLAYER_FIELD_DILATION_RATE(), dilation);
+        }
+        List<Integer> subsampleList = new ArrayList<>();
+        subsampleList.add(STRIDE[0]);
+        subsampleList.add(STRIDE[1]);
+        config.put(conf.getLAYER_FIELD_CONVOLUTION_STRIDES(), subsampleList);
+        config.put(conf.getLAYER_FIELD_NB_FILTER(), N_OUT);
+        config.put(conf.getLAYER_FIELD_BORDER_MODE(), BORDER_MODE_VALID);
+        layerConfig.put(conf.getLAYER_FIELD_CONFIG(), config);
+        layerConfig.put(conf.getLAYER_FIELD_KERAS_VERSION(), kerasVersion);
+
+
+        ConvolutionLayer layer = new KerasConvolution2D(layerConfig).getConvolution2DLayer();
+        assertEquals(ACTIVATION_DL4J, layer.getActivationFn().toString());
+        assertEquals(LAYER_NAME, layer.getLayerName());
+        assertEquals(INIT_DL4J, layer.getWeightInit());
+        assertEquals(L1_REGULARIZATION, layer.getL1(), 0.0);
+        assertEquals(L2_REGULARIZATION, layer.getL2(), 0.0);
+        assertEquals(DROPOUT_DL4J, layer.getDropOut(), 0.0);
+        assertArrayEquals(KERNEL_SIZE, layer.getKernelSize());
+        assertArrayEquals(STRIDE, layer.getStride());
+        assertEquals(N_OUT, layer.getNOut());
+        assertEquals(ConvolutionMode.Truncate, layer.getConvolutionMode());
+        assertArrayEquals(VALID_PADDING, layer.getPadding());
+        if (withDilation) {
+            assertEquals(DILATION[0], layer.getDilation()[0]);
+            assertEquals(DILATION[1], layer.getDilation()[1]);
+        }
+    }
+}
